Fixed compiler warnings (#41)
* Update hbencode.c

* Update gtwvw.c

* Update gtwvg.c

* Update inetssl.c

* Update ssl.c

* Update sslctx.c

* Update compile.mak

* Add files via upload

* Add files via upload

* Update ChangeLog

+++ contrib/hbssl/ssl.c
@@ -1276,7 +1276,7 @@ HB_FUNC( SSL_GET_OPTIONS )
       SSL * ssl = hb_SSL_par( 1 );
 
       if( ssl )
-         hb_retnl( SSL_get_options( ssl ) );
+         hb_retnl( ( LONG ) SSL_get_options( ssl ) );
    }
    else
       hb_errRT_BASE( EG_ARG, 2010, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
